refactor: rename PdfViewer to FormEditorPdfViewer, use BasePdfViewer
Refactored to use BasePdfViewer internally. pdf-viewer.context now uses
usePdfDocumentFromFile from shared hooks instead of manual FileReader + pdfjs.
Updated FormEditorTab import.
